# Add CCF codec implementing CCF Specification (RC1)