… .golangci.yml
… encoding/ccf/bench_test.go
… encoding/ccf/cache.go
… encoding/ccf/ccf_test.go
… encoding/ccf/ccf_type_id_test.go
… encoding/ccf/decode_typedef.go
… encoding/ccf/encode.go
… encoding/ccf/encode_type.go
… encoding/ccf/encode_typedef.go
… encoding/ccf/sort.go
… encoding/ccf/traverse_value.go
… types.go
… types_test.go
… values.go