## Description

Updates #2157 

[Cadence Compact Format (CCF)](https://github.com/fxamacker/ccf_draft) is a data format designed for compact, efficient, and deterministic encoding of Cadence external values.

CCF obsoletes [JSON-Cadence Data Interchange Format](https://developers.flow.com/cadence/json-cadence-spec) (JCDIF) for use cases that do not require JSON.

Unlike JCDIF, etc. CCF Specifications explicitly defines requirements for:
- well-formed encodings
- valid encodings
- deterministic encodings

CCF is hybrid data format. CCF-based messages can be:
- fully self-describing or
- partially self-describing.

Both CCF modes are more compact than JSON-based messages. CCF-based protocols can send Cadence metadata just once for all messages of that type. Malformed data can be detected without Cadence metadata and without creating Cadence objects.

CCF codec implements [CCF Specification (RC1)](https://github.com/fxamacker/ccf_draft/commit/4e8f42db29925bd15481301729eca1b52852dcb4), which is temporarily at [github.com/fxamacker/ccf_draft](https://github.com/fxamacker/ccf_draft). The CCF specs
will be hosted under [github.com/onflow](https://github.com/onflow/cadence) after it is updated, cleaned up, and reformatted.

### API

CCF codec provides the same API as the existing JSON-Cadence Data Interchange codec.

Given this, integration effort to replace the old codec should be minimal.

### PRELIMINARY COMPARISONS

Comparisons used 48,309 events from a single mainnet transaction.

There were 9 event types. To simplify benchmark code, the first event's value in each of the 9 event types was used.

CCF's partially self-describing mode (aka "detached" mode) would be even smaller than this (e.g. maybe less than 1/4 the size of JSON when Flow eventually supports detached mode).

#### SIZE COMPARISON

```
Encoding | Num Events | Encoded size | Comments
-------- | ---------- | ------------ | --------
JSON     |     48,309 |   13,858,836 | JSON-Cadence Data Interchange
CCF      |     48,309 |    6,159,931 | CCF in fully self-describing mode
```

#### CCF SPEED AND MEMORY COMPARISONS

This isn't apples to apples comparison.   JSON data isn't sorted, etc.
- CCF encoder sorts data to encode event data deterministically.
- CCF decoder also verifies event data is sorted, well-formed, and valid.
- CCF encoding is less than 1/2 size of JSON-Cadence Data Interchange.

```
$ benchstat bench_json_events_48k.log bench_ccf_events_48k.log 
goos: linux
goarch: amd64
pkg: github.com/onflow/cadence/encoding/ccf
cpu: 13th Gen Intel(R) Core(TM) i5-13600K
                     │ bench_json_events_48k.log │      bench_ccf_events_48k.log       │
                     │          sec/op           │   sec/op     vs base                │
EncodeBatchEvents-20                 96.61m ± 4%   70.73m ± 3%  -26.79% (p=0.000 n=10)
DecodeBatchEvents-20                 647.7m ± 3%   157.5m ± 3%  -75.68% (p=0.000 n=10)
geomean                              250.1m        105.5m       -57.81%

                     │ bench_json_events_48k.log │       bench_ccf_events_48k.log       │
                     │           B/op            │     B/op      vs base                │
EncodeBatchEvents-20                32.45Mi ± 0%   25.82Mi ± 0%  -20.45% (p=0.000 n=10)
DecodeBatchEvents-20               234.97Mi ± 0%   56.16Mi ± 0%  -76.10% (p=0.000 n=10)
geomean                             87.32Mi        38.08Mi       -56.39%

                     │ bench_json_events_48k.log │      bench_ccf_events_48k.log       │
                     │         allocs/op         │  allocs/op   vs base                │
EncodeBatchEvents-20                 756.6k ± 0%   370.4k ± 0%  -51.05% (p=0.000 n=10)
DecodeBatchEvents-20                 4.746M ± 0%   1.288M ± 0%  -72.86% (p=0.000 n=10)
geomean                              1.895M        690.7k       -63.55%
```

Benchmarked using Go 1.19.6, linux_amd64, i5-13600k.  Results are subject to change because CCF codec reference implementation in Go has not yet been reviewed.

### NEXT STEPS

:point_right: After this PR is merged, there's additional work that should be completed before using or deploying this codec.

- [x] Add more tests to  handle edge cases (as time allows).  Currently, `go test -cover` reports 77.3% for CCF codec.

- [x] Add and run fuzz tests.  It would be very unusual for fuzzing to not find any problems with a brand new codec for a new data format.  I think Cadence Team is best qualified for this task (i.e. expert in Cadence types & values).

- [ ] Integration with FVM to use CCF codec for events.

- [ ] Integration work related to converting CCF encodings to JSON-Cadence Data Interchange.

- [ ] Add integration tests in relevant projects.

EDIT: updated benchmarks to match commit f911063ca19956976ac3d9f15bc5a35bfaa0cf91

______

<!-- Complete: -->

- [x] Targeted PR against `master` branch
- [x] Linked to Github issue with discussion and accepted design OR link to spec that describes this work
- [x] Code follows the [standards mentioned here](https://github.com/onflow/cadence/blob/master/CONTRIBUTING.md#styleguides)
- [ ] Updated relevant documentation 
- [x] Re-reviewed `Files changed` in the Github PR explorer
- [x] Added appropriate labels