@fxamacker 2023-03-17
> Noticed that we may need to add support for the newly added attachment type and value.

I agree. There is an issue #2379 to add support for recently added types (e.g. `AttachmentType`).  I plan to add support in a followup PR (this one is already very large).

@SupunS, is there a way for me to get automatically notified when external types & values gets added or modified?