@SupunS 2023-03-17
oh okay, didn't know there was an issue already 👍 I Just ran onto this while fuzzing. Agree that it could fit into a separate PR.

> is there a way for me to get automatically notified when external types & values gets added or modified?

Unfortunately no. Adding new types / modifying happens very rarely, and would be even less frequent (ideally zero) once stable-cadence is released. Currently, we manually ensure all components are updated when the type is added/modified. e.g: One who add/modify a type would create an issue to update the encoder (or even include the changes, if feasible).
But on this occasion, I guess it's a temporary situation where the encoder and attachments feature were developed parallelly, and hence run into this problem where two components are not in sync.