@fxamacker 2023-04-13
@turbolent @SupunS Thanks for reviews, fuzzing, and meetings!  After merging this, I'll open followup PR to add API to allow some default CBOR limits to be overridden (e.g. `NewDecoderWithOptions()`).

I spotted an edge case where it was possible for encoder to needlessly encode redundant data, related to reference type.  

Can you take a look at commit 51552f153326be5a493682fcd68ae381d9899013 as a sanity check before I merge this PR?  :pray: